# Conflicts:
#	internal/internal_flags.go
#	internal/internal_workflow_testsuite_test.go

+++ internal/internal_task_handlers_interfaces_test.go
@@ -196,7 +196,7 @@ func (s *PollLayerInterfacesTestSuite) TestGetNextCommandsSdkFlags() {
 			ScheduledEventId: 2,
 			StartedEventId:   3,
 			SdkMetadata: &sdk.WorkflowTaskCompletedMetadata{
-				LangUsedFlags: []uint32{SDKFlagLimitChangeVersionSASize},
+				LangUsedFlags: []uint32{uint32(SDKFlagLimitChangeVersionSASize)},
 				SdkName:       SDKName,
 				SdkVersion:    "1.0",
 			},
